Revert "Move Workers to data layer"
This reverts commit 8260885dbe7ae2af4a8d3e43bf8a54213b0b8c6e.
